+++ main.go
@@ -5,5 +5,5 @@ import "github.com/nlink-jp/stail/internal/cmd"
 var version = "dev"
 
 func main() {
-	cmd.Execute()
+	cmd.Execute(version)
 }